ai-agent/requirements.txt
strands-agents
mcp
boto3